refactor(logic operators): Migrate to typescript. Add operator type enum.
