+++ src/js/Enums/LogicOperatorType.ts
@@ -1,0 +1,9 @@
+/**
+ * Type of logic operator.
+ */
+enum LogicOperatorType {
+	And = 'And',
+	Or = 'Or',
+}
+
+export default LogicOperatorType;
